validate

+++ LSAPI/Controllers/BatchController.cs
@@ -2,10 +2,13 @@
 using LocalShipper.Service.DTOs.Response;
 using LocalShipper.Service.Services.Implement;
 using LocalShipper.Service.Services.Interface;
+using MailKit;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Org.BouncyCastle.Asn1.Ocsp;
 using System;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace LSAPI.Controllers
@@ -26,6 +29,23 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if (pageNumber.HasValue && pageNumber < 0)
+                {
+                    return BadRequest("pageNumber phải là số dương");
+                }
+
+                if (pageSize.HasValue && pageSize < 0)
+                {
+                    return BadRequest("pageSize phải là số dương");
+                }
+                if (id < 0 )
+                {
+                    return BadRequest("id không hợp lệ");
+                }
+
+
+
+
                 var rs = await _batchService.GetBatch(id, batchName, pageNumber, pageSize);
                 return Ok(rs);
             }
@@ -41,6 +61,24 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if(request.StoreId <= 0 )
+                {
+                    return BadRequest("storeId phải là 1 số dương");
+                }
+
+                var regex = new Regex("^[a-zA-Z0-9 ]+$"); 
+                if (!regex.IsMatch(request.BatchName))
+                {
+                    return BadRequest("batchName không được chứa ký tự đặc biệt");
+                }
+                if (!regex.IsMatch(request.BatchDescription))
+                {
+                    return BadRequest("batchDescription không được chứa ký tự đặc biệt");
+                }
+                if (request.Status <= 0)
+                {
+                    return BadRequest("status phải phải là 1 số dương");
+                }
                 var rs = await _batchService.CreateBatch(request);
                 return Ok(rs);
             }
@@ -57,6 +95,32 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
+                if (id <= 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
+                if (batchRequest.StoreId <= 0)
+                {
+                    return BadRequest("storeId phải là 1 số dương");
+                }
+
+                var regex = new Regex("^[a-zA-Z0-9 ]+$");
+                if (!regex.IsMatch(batchRequest.BatchName))
+                {
+                    return BadRequest("batchName không được chứa ký tự đặc biệt");
+                }
+                if (!regex.IsMatch(batchRequest.BatchDescription))
+                {
+                    return BadRequest("batchDescription không được chứa ký tự đặc biệt");
+                }
+                if(batchRequest.Status <= 0)
+                {
+                    return BadRequest("status phải phải là 1 số dương");
+                }
                 var rs = await _batchService.UpdateBatch(id, batchRequest);
                 return Ok(rs);
             }
@@ -67,17 +131,19 @@ namespace LSAPI.Controllers
 
         }
 
-
-        /// <summary>
-        /// Xóa một batch bằng cách cập nhật trạng thái thành 4 (DELETE).
-        /// </summary>
-        /// <param name="id">ID của batch cần xóa.</param>
-        /// <returns>Thông tin của batch đã bị xóa.</returns>
         [HttpDelete()]
         public async Task<ActionResult<BatchResponse>> DeleteBatch(int id)
         {
             try
             {
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
+                if (id <= 0) 
+                {
+                    return BadRequest("id phải là số dương");
+                }
                 var rs = await _batchService.DeleteBatch(id);
                 return Ok(rs);
             }
