+++ LSAPI/Controllers/PackageController.cs
@@ -5,8 +5,10 @@ using LocalShipper.Service.Services.Implement;
 using LocalShipper.Service.Services.Interface;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Org.BouncyCastle.Asn1.X509;
 using System;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace LSAPI.Controllers
@@ -25,13 +27,46 @@ namespace LSAPI.Controllers
         [HttpGet()]
 
         public async Task<ActionResult<List<PackageResponse>>> GetPackage(int batchId, int id,
-            int status, int actionId, int typeId,int storeId, string customerName, string customerAddress,
-            string customerPhome, string custommerEmail, decimal totalPrice, int? pageNumber, int? pageSize)
+            int status, int actionId, int typeId, int storeId, string customerName,
+             int? pageNumber, int? pageSize)
         {
             try
             {
-                var package = await _packageService.GetPackage(batchId, id, status, actionId, typeId, storeId, 
-                    customerName, customerAddress, customerPhome, custommerEmail, totalPrice, pageNumber, pageSize);
+                if (pageNumber.HasValue && pageNumber < 0)
+                {
+                    return BadRequest("pageNumber phải là số dương");
+                }
+
+                if (pageSize.HasValue && pageSize < 0)
+                {
+                    return BadRequest("pageSize phải là số dương");
+                }
+                if (batchId < 0)
+                {
+                    return BadRequest("batchId phải là số dương");
+                }
+                if (id < 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
+                if (status < 0)
+                {
+                    return BadRequest("status phải là số dương");
+                }
+                if (actionId < 0)
+                {
+                    return BadRequest("actionId phải là số dương");
+                }
+                if (typeId < 0)
+                {
+                    return BadRequest("typeOd phải là số dương");
+                }
+                if (storeId < 0)
+                {
+                    return BadRequest("storeId phải là số dương");
+                }
+                var package = await _packageService.GetPackage(batchId, id, status, actionId, typeId, storeId,
+                    customerName, pageNumber, pageSize);
                 return Ok(package);
 
             }
@@ -46,6 +81,57 @@ namespace LSAPI.Controllers
         {
             try
             {
+
+
+                if (request.StoreId < 0)
+                {
+                    return BadRequest("storeId phải là số dương");
+                }
+                if (request.Capacity < 0)
+                {
+                    return BadRequest("Capacity phải là số dương");
+                }
+                if (request.PackageWidth < 0)
+                {
+                    return BadRequest("PackageWidth phải là số dương");
+                }
+                if (request.PackageWeight < 0)
+                {
+                    return BadRequest("PackageWeight phải là số dương");
+                }
+                if (request.PackageHeight < 0)
+                {
+                    return BadRequest("PackageHeight phải là số dương");
+                }
+                if (request.PackageLength < 0)
+                {
+                    return BadRequest("PackageLength phải là số dương");
+                }
+                var regex = new Regex(@"^\w+@gmail\.com$");
+                if (!regex.IsMatch(request.CustomerEmail))
+                {
+                    return BadRequest("Email phải có địa chỉ tên miền @gmail.com");
+                }
+                if (request.PackagePrice < 0)
+                {
+                    return BadRequest("PackagePrice phải là số dương");
+                }
+                if (request.DistancePrice < 0)
+                {
+                    return BadRequest("DistancePrice phải là số dương");
+                }
+                if (request.SubtotalPrice < 0)
+                {
+                    return BadRequest("SubtotalPrice phải là số dương");
+                }
+                if (request.ActionId < 0)
+                {
+                    return BadRequest("ActionId phải là số dương");
+                }
+                if (request.TypeId < 0)
+                {
+                    return BadRequest("TypeId phải là số dương");
+                }
                 var rs = await _packageService.CreatePackage(request);
                 return Ok(rs);
             }
@@ -60,10 +146,62 @@ namespace LSAPI.Controllers
         {
             try
             {
-                if (request.StoreId.HasValue)
+                if (id == 0)
                 {
-                    request.StoreId = null;
+                    return BadRequest("làm ơn hãy nhập id");
                 }
+                if (id < 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
+
+                if (request.Capacity <= 0)
+                {
+                    return BadRequest("Capacity phải là số dương");
+                }
+                if (request.PackageWidth <= 0)
+                {
+                    return BadRequest("PackageWidth phải là số dương");
+                }
+                if (request.PackageWeight <= 0)
+                {
+                    return BadRequest("PackageWeight phải là số dương");
+                }
+                if (request.PackageHeight <= 0)
+                {
+                    return BadRequest("PackageHeight phải là số dương");
+                }
+                if (request.PackageLength <= 0)
+                {
+                    return BadRequest("PackageLength phải là số dương");
+                }
+                var regex = new Regex(@"^\w+@gmail\.com$");
+                if (!regex.IsMatch(request.CustomerEmail))
+                {
+                    return BadRequest("Email phải có địa chỉ tên miền @gmail.com");
+                }
+                if (request.PackagePrice < 0)
+                {
+                    return BadRequest("PackagePrice phải là số dương");
+                }
+                if (request.DistancePrice < 0)
+                {
+                    return BadRequest("DistancePrice phải là số dương");
+                }
+                if (request.SubtotalPrice < 0)
+                {
+                    return BadRequest("SubtotalPrice phải là số dương");
+                }
+                if (request.ActionId <= 0)
+                {
+                    return BadRequest("ActionId phải là số dương");
+                }
+                if (request.TypeId <=0)
+                {
+                    return BadRequest("TypeId phải là số dương");
+                }
+
+
                 var rs = await _packageService.UpdatePackage(id, request);
                 return Ok(rs);
             }
@@ -75,10 +213,23 @@ namespace LSAPI.Controllers
         }
 
         [HttpPut("status")]
-        public async Task<ActionResult<PackageResponse>> PutStatusPackage(int id,  PackageStatusEnum status)
+        public async Task<ActionResult<PackageResponse>> PutStatusPackage(int id, PackageStatusEnum status)
         {
             try
             {
+
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
+                if (id < 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
+                if(status <= 0)
+                {
+                    return BadRequest("status phải là số dương");
+                }
                 var rs = await _packageService.UpdateStatusPackage(id, status);
                 return Ok(rs);
             }
@@ -94,6 +245,14 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
+                if (id <= 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
                 var rs = await _packageService.DeletePackage(id);
                 return Ok(rs);
             }
